+++ examples/autoreload/src/main.rs
@@ -4,7 +4,7 @@ use std::thread;
 use std::time::Duration;
 
 use minijinja::context;
-use minijinja::{Environment, Source};
+use minijinja::{path_loader, Environment};
 use minijinja_autoreload::AutoReloader;
 
 fn main() {
@@ -14,15 +14,14 @@ fn main() {
     // The closure is invoked every time the environment is outdated to
     // recreate it.
     let reloader = AutoReloader::new(move |notifier| {
-        let mut env = Environment::new();
         let template_path = PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("templates");
+        let mut env = Environment::new();
+        env.set_loader(path_loader(&template_path));
 
         // if watch_path is never called, no fs watcher is created
         if !disable_autoreload {
             notifier.watch_path(&template_path, true);
         }
-
-        env.set_source(Source::from_path(&template_path));
         Ok(env)
     });